More code review comments addressed.
