+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/sqlbuilder/ReportQueryProvider.java
@@ -1,0 +1,17 @@
+package org.opentestsystem.rdw.olap.sqlbuilder;
+
+import org.opentestsystem.rdw.reporting.common.model.ReportQuery;
+
+
+/**
+ * Implementation of this interface is responsible for converting a {@link ReportQuery} to a sql query string
+ */
+public interface ReportQueryProvider {
+    /**
+     * Converts a given {@link ReportQuery} to a sql query string
+     *
+     * @param query {@link ReportQuery}
+     * @return a sql query string
+     */
+    String toSqlQueryString(ReportQuery query);
+}
