+++ aggregate-service/src/main/resources/application.sql.yml
@@ -28,54 +28,6 @@ sql:
   subject:
     findAll: select id, code from subject
 
-
-  # a tokenized version of the SQL query for the custom aggregate reports
-  # the @{template.XXXX} tokens are replaced with the appropriate queries at run time
-  customAggregate: >-
-      CASE WHEN dim.organization_id = -1 THEN '${app.state.name}' else dim.organization_name END,
-      CASE WHEN dim.organization_id = -1 THEN null else dim.organization_id END,
-      dim.organization_type,
-      dim.grade_id as asmt_grade_id,
-      dim.school_year,
-      dim.subject_id,
-      dim.dimension_code,
-      dim.dimension,
-      dim.asmt_id,
-      measure.score,
-      measure.std_err,
-      measure.level1,
-      measure.level2,
-      measure.level3,
-      measure.level4
-      FROM (@{template.customAggregateAllPermutations}) dim
-        LEFT JOIN (@{template.customAggregate}) measure
-      ON measure.asmt_grade_id = dim.grade_id
-      AND measure.subject_id = dim.subject_id
-      AND dim.school_year = measure.school_year
-      AND dim.organization_id = measure.organization_id
-      AND measure.dimension_id = dim.dimension_id
-      AND measure.asmt_id = dim.asmt_id
-      AND measure.dimension_id = dim.dimension_id
-
- # unlike all other dimensions we are not back-filling missing enrolled grades
-  customAggregateByEnrolledGrade: >-
-     CASE WHEN organization_id = -1 THEN '${app.state.name}' else organization_name END,
-     CASE WHEN organization_id = -1 THEN null else organization_id END,
-     organization_type,
-     asmt_grade_id,
-     school_year,
-     subject_id,
-     dimension_code,
-     dimension,
-     asmt_id,
-     score,
-     std_err,
-     level1,
-     level2,
-     level3,
-     level4
-     FROM (@{template.customAggregate}) a
-
   # SQL does not support an Aggregate Product, but this seems to be a recommended way
   estimateReportRowCount: >-
    SELECT round(exp(sum(ln(count)))) as count from (
