+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/repository/impl/JdbcCustomAggregateReportRowRepositoryRST.java
@@ -39,7 +39,7 @@ import static org.springframework.test.jdbc.JdbcTestUtils.countRowsInTable;
 
 @SuppressWarnings({"SqlNoDataSourceInspection", "SqlResolve"})
 
-@Import(JdbcCustomAggregateReportRowRepository.class)
+@Import({JdbcCustomAggregateReportRowRepository.class})
 @Sql(scripts = {"classpath:OlapEntitiesSetup.sql"})
 public class JdbcCustomAggregateReportRowRepositoryRST extends RepositoryBackedIT {
