+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/repository/impl/RepositoryBackedIT.java
@@ -2,6 +2,7 @@ package org.opentestsystem.rdw.olap.repository.impl;
 
 import org.junit.runner.RunWith;
 import org.opentestsystem.rdw.olap.repository.DataSourceConfiguration;
+import org.opentestsystem.rdw.olap.sqlbuilder.CustomAggregateQueryProvider;
 import org.opentestsystem.rdw.reporting.common.sqlbuilder.QueryProvider;
 import org.opentestsystem.rdw.reporting.common.sqlbuilder.SqlBuilderConfig;
 import org.opentestsystem.rdw.utils.YamlPropertiesConfigurator;
@@ -16,7 +17,8 @@ import org.springframework.transaction.annotation.Transactional;
         DataSourceConfiguration.class,
         YamlPropertiesConfigurator.class,
         QueryProvider.class,
-        SqlBuilderConfig.class
+        SqlBuilderConfig.class,
+        CustomAggregateQueryProvider.class
 })
 @ContextConfiguration(classes = {TestDataSourceConfiguration.class})
 @Transactional(transactionManager = "txManager")
